Wire auto-run into ingest flow and remove old trigger infrastructure
Hook sample completeness check into manifest entry verification in
process_ingest_event. On verified: queue auto-runs. On checksum
mismatch: cancel waiting runs. Replace three old trigger background
loops with single auto_run_launch_loop. Delete trigger_service,
pipeline_triggers API, and trigger schemas. Move BudgetCheckResult
to budget_service. Update affected tests.
